+++ sql/updates/latest-dev.sql
@@ -1,0 +1,4 @@
+DROP FUNCTION _timescaledb_internal.ping_data_node(NAME);
+
+CREATE FUNCTION _timescaledb_internal.ping_data_node(node_name NAME, timeout INTERVAL = NULL) RETURNS BOOLEAN
+AS '@MODULE_PATHNAME@', 'ts_data_node_ping' LANGUAGE C VOLATILE;